wip make httpserver.close_all_conns() close websockets too
Just a test for now; this is tricky.

I tried doing this by making detach() not set finish_future so the
http layers consider the connection open for as long as
_receive_frame_loop (and therefore handler.get()) stay pending. This
required an unsightly amount of plumbing and special-casing (never finished) to for
example prevent RequestHandler.finish from writing anything to the
socket but still signaling to the h1conn that it's done.

I think the better approach is for detach() to take a callback. This
callback would either be plumbed up to the HTTPServer to take the
h1conn's place in the active connection set (difficult with current
organization) or the h1conn would stay in its penultimate state until
closed, at which point it would invoke this callback.